// domain/types/type.ts
export enum CategoryType {
  Movie = "movie",
  Tv = "tv",
}

export enum MovieType {
  Popular = "popular",
  TopRated = "top_rated",
  Similar = "similar",
  Upcoming = "upcoming",
}

export enum TvType {
  Popular = "popular",
  TopRated = "top_rated",
  Similar = "similar",
  OnTheAir = "on_the_air",
}